+++ lib/neoprism_core.dart
@@ -1,4 +1,4 @@
-library;
+library neoprism_core;
 
 // Export components
 export 'src/components/abstract/neoprism_component.dart';
@@ -6,6 +6,7 @@ export 'src/components/buttons/neo_button.dart';
 export 'src/components/badges/neo_badge.dart';
 export 'src/components/input/neo_input.dart';
 export 'src/components/checkbox/neo_checkbox.dart';
+export 'src/components/alert/alert_dialog/neo_alert_dialog.dart';
 
 // Export theme
 export 'src/theme/theme.dart';